Convert to use jh.JSONHandler
including:

    - remove sendError and sendResp
    - split up api_handlers.go into individual parts
